--- include/sliding_window/argument_parsing/shared.hpp
@@ -1,26 +1,0 @@
-#pragma once
-
-#include <sliding_window/argument_parsing/validators.hpp>
-#include <sliding_window/shared.hpp>
-
-namespace sliding_window::app
-{
-
-void init_shared_meta(seqan3::argument_parser & parser);
-void try_parsing(seqan3::argument_parser & parser);
-
-template <typename arguments_t>
-void init_shared_options(seqan3::argument_parser & parser, arguments_t & arguments)
-{
-    static_assert(std::same_as<arguments_t, build_arguments> || std::same_as<arguments_t, search_arguments>);
-
-    // TODO: split cannot be run multi-threaded
-    parser.add_option(arguments.threads,
-                      '\0',
-                      "threads",
-                      "Choose the number of threads.",
-                      seqan3::option_spec::standard,
-                      positive_integer_validator{});
-}
-
-} // namespace sliding_window::app
