--- include/sliding_window/argument_parsing/top_level.hpp
@@ -1,10 +1,0 @@
-#pragma once
-
-#include <sliding_window/argument_parsing/shared.hpp>
-
-namespace sliding_window::app
-{
-
-void init_top_level_parser(seqan3::argument_parser & parser);
-
-} // namespace sliding_window::app
