--- include/sliding_window/search/search.hpp
@@ -1,10 +1,0 @@
-#pragma once
-
-#include <sliding_window/shared.hpp>
-
-namespace sliding_window::app
-{
-
-void sliding_window_search(search_arguments const & arguments);
-
-} // namespace sliding_window::app
