--- include/sliding_window/split/split.hpp
@@ -1,10 +1,0 @@
-#pragma once
-
-#include <sliding_window/shared.hpp>
-
-namespace sliding_window::app
-{
-
-void sliding_window_split(split_arguments const & arguments);
-
-} // namespace sliding_window::app
