+++ include/valik/argument_parsing/top_level.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <valik/argument_parsing/shared.hpp>
+
+namespace valik::app
+{
+
+void init_top_level_parser(seqan3::argument_parser & parser);
+
+} // namespace valik::app
