+++ include/valik/search/search.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <valik/shared.hpp>
+
+namespace valik::app
+{
+
+void valik_search(search_arguments const & arguments);
+
+} // namespace valik::app
