+++ include/valik/split/split.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <valik/shared.hpp>
+
+namespace valik::app
+{
+
+void valik_split(split_arguments const & arguments);
+
+} // namespace valik::app
